pyproject.toml
[project]
name = "la-bench"
version = "0.1.0"
description = "LA-Bench 2025: 実験手順生成タスク"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "trafilatura>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
